third commit
made the homepage an actual homepage. added a site for every genre. added links to each site in the header. changed the look of the pages with the movies a little bit. think i made the code better as well. removed nesting in css because it was giving me errors even though it worked, i dont know why. added contact information at the bottom of the page. changed font.