fourth commit
changed the code for the header to work for mobile. finished the content for romance, drama and comedy. made the content in the contact div centered and added an a tag for my mail. changed the homepage into selection for a genre and removed genre selection from the header(only on the homepage). this made a lot more sense and looked way better in my opinion.